add unit tests

+++ src/main/java/com/connex/car/insurance/quote/dto/response/QuoteData.java
@@ -1,6 +1,6 @@
 package com.connex.car.insurance.quote.dto.response;
 
-public record QuoteData(String year, Integer carCost, Boolean business, Integer kilometres, Integer accidents,
-        Integer claims, Integer personAge, Integer yearsInsuranceHistory) {
+public record QuoteData(String year, Double carCost, Boolean business, Integer kilometres, Integer accidents,
+                Integer claims, Integer personAge, Integer yearsInsuranceHistory) {
 
 }
